requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
